src/requirements.txt
llama-index==0.10.34
llama-index-postprocessor-sbert-rerank==0.1.2
openai==1.23.6
pandas==2.2.1
PyYAML==6.0.1
tqdm==4.66.2
streamlit==1.34.0
ragas==0.1.9
markitdown==0.3.0 # Or whatever version you have
sentence-transformers==2.7.0
torch==2.2.2 # Or your specific torch version
markitdown[pdf]
torch
sentence-transformers
tf-keras